Move contents of aiida.utils.logger to aiida.common.log
The module aiida.common.log now provides all the logging
related concepts within AiiDA.
