+++ aiida/backends/djsite/utils.py
@@ -10,7 +10,7 @@
 
 import os
 import django
-from aiida.utils.logger import get_dblogger_extra
+from aiida.common.log import get_dblogger_extra
 
 
 def load_dbenv(process=None, profile=None):
