+++ aiida/backends/sqlalchemy/cmdline.py
@@ -8,7 +8,7 @@
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
 
-from aiida.utils.logger import get_dblogger_extra
+from aiida.common.log import get_dblogger_extra
