+++ aiida/daemon/execmanager.py
@@ -36,7 +36,7 @@ def update_running_calcs_status(authinfo):
     """
     from aiida.orm import JobCalculation, Computer
     from aiida.scheduler.datastructures import JobInfo
-    from aiida.utils.logger import get_dblogger_extra
+    from aiida.common.log import get_dblogger_extra
     from aiida.backends.utils import QueryFactory
     
     if not authinfo.enabled:
@@ -267,7 +267,7 @@ def submit_jobs():
     Submit all jobs in the TOSUBMIT state.
     """
     from aiida.orm import JobCalculation, Computer, User
-    from aiida.utils.logger import get_dblogger_extra
+    from aiida.common.log import get_dblogger_extra
     from aiida.backends.utils import get_authinfo, QueryFactory
 
 
@@ -336,7 +336,7 @@ def submit_jobs_with_authinfo(authinfo):
     to user and machine as defined in the 'dbauthinfo' table.
     """
     from aiida.orm import JobCalculation
-    from aiida.utils.logger import get_dblogger_extra
+    from aiida.common.log import get_dblogger_extra
 
     from aiida.backends.utils import QueryFactory
 
@@ -384,7 +384,7 @@ def submit_jobs_with_authinfo(authinfo):
         # because any other exception is caught and skipped above
         except Exception as e:
             import traceback
-            from aiida.utils.logger import get_dblogger_extra
+            from aiida.common.log import get_dblogger_extra
 
             for calc in calcs_to_inquire:
                 logger_extra = get_dblogger_extra(calc)
@@ -423,7 +423,7 @@ def submit_calc(calc, authinfo, transport=None):
     from aiida.common.exceptions import (
         InputValidationError)
     from aiida.orm.data.remote import RemoteData
-    from aiida.utils.logger import get_dblogger_extra
+    from aiida.common.log import get_dblogger_extra
 
     if not authinfo.enabled:
         return
@@ -664,7 +664,7 @@ def retrieve_computed_for_authinfo(authinfo):
     from aiida.orm import JobCalculation
     from aiida.common.folders import SandboxFolder
     from aiida.orm.data.folder import FolderData
-    from aiida.utils.logger import get_dblogger_extra
+    from aiida.common.log import get_dblogger_extra
     from aiida.orm import DataFactory
 
     from aiida.backends.utils import QueryFactory
