+++ aiida/orm/implementation/general/calculation/__init__.py
@@ -129,7 +129,7 @@ class AbstractCalculation(SealableWithUpdatableAttributes):
           the 'extra' embedded
         """
         import logging
-        from aiida.utils.logger import get_dblogger_extra
+        from aiida.common.log import get_dblogger_extra
 
         return logging.LoggerAdapter(
             logger=self._logger, extra=get_dblogger_extra(self))
