+++ aiida/orm/implementation/general/workflow.py
@@ -26,7 +26,7 @@ from aiida.orm.implementation.calculation import JobCalculation
 from aiida.backends.utils import get_automatic_user
 
 from aiida.utils import timezone
-from aiida.utils.logger import get_dblogger_extra
+from aiida.common.log import get_dblogger_extra
 from aiida.common.utils import abstractclassmethod
 
 logger = aiidalogger.getChild('Workflow')
