+++ aiida/parsers/parser.py
@@ -38,7 +38,7 @@ class Parser(object):
         extras of the calculation
         """
         import logging
-        from aiida.utils.logger import get_dblogger_extra
+        from aiida.common.log import get_dblogger_extra
 
         return logging.LoggerAdapter(logger=self._logger, extra=get_dblogger_extra(self._calc))
